Framework usage. End-to-end implementation of business rules for Entity Framework
The required database object is created in the Entity layer. The created object must bear the IEntity signature, which is a business rule. Optional Mapping is applied. DbSet is created in the Context class. The interface of the related object is created and inherited. The data access class of the relevant object implements only the interface in which the operations occur and the class in which the operations occur.